Increase utility of styleMap in text manipulation
By putting the first capture group in as the magic number in the
focus offset, this will let us remove the keyCharsCount from the
styleMap and make the main code more readable.

This will also allow multiple selectors of different lengths in
the regex. For instance, a single backtick and triple backticks will now
both result in the same outcome without having to do some weird
workaround re: character counting.