Added image drag and drop
Added the ability for the user to drag any image media (PNG, JPEG, BMP,
ect...) into the editor area.

One issue I was having in doing this was that if you drop directly into
the draftjs DOM element (by giving it an event listener and detecting a
drop), draftjs breaks. Not sure what the root cause is, but the blocks
in the contentState stop updating after the drop and no markdown is
rendered. To get around this, I added a textArea that pops up only when
something is dragged into the editor area. That way, the listener is not
attached to the draftjs editor and everything keeps working.

Current known issues:

	- Occasionally, entering a new line of text will copy the image.
	  Have not yet looked into why this happens.
	- A drag without a drop results in the editor getting hidden
	  behind a transparent textArea. Need to add logic to determine
	  if the drag was resolved (dropped or dragged out of the
	  editor) to fix this. Should be simple.