Renamed/Enabled renderInlineStyles
Currently, block and inline style renderings are working as expected
with the exception that an extra line is added upon each backspace. This
is because both render functions perform a split-block at the end.

Need to reconcile the ending of both of the render functions.

Other that that, this should be functional for block stylings.