Put styling shortcuts behind flag.
Added call to flaginator-9000 to turn key styling shortcuts on based on
flag. To flip this flag on or off, go to dev.dthomas.io and log in with
email: notes@notesapp.com & password: notesnotes1.

We clearly need to refine how we handle flags, but this is just a start.